refactor(api): extract the RPC-proxy, staging, and entity-handler clusters (finish #1763)
Move the final three handler clusters out of workers/api.mjs into their own
modules under workers/request-handlers/, following the analytics.mjs extraction
pattern from #1787 (explicit injected deps, no shared mutable module-globals
across files, no import cycles). Pure refactor — zero behavior change.

- staging.mjs: the four loadStaged* loaders (neurons/events/blocks/extrinsics)
  plus the HMAC signing/validation machinery they alone use. No injected deps
  (all leaf imports). api.mjs re-exports the loaders for the cron + tests.
- entities.mjs: the single-entity D1 lookups (subnet metagraph/validators,
  per-UID neuron + history, per-subnet history, account summary/events/subnets,
  block + extrinsic feeds and detail). Pulls d1All/d1Runner + the query guards
  directly from the sibling analytics.mjs (no cycle).
- rpc-proxy.mjs: the /rpc/v1 proxy + B3 usage analytics + GraphQL rate-limit
  guard + surface-verify probe, with the RPC_HEALTH circuit-breaker co-located
  with its sole reader/writer (proxyWithFailover / orderSafeRpcEndpoints), the
  pool-artifact memo, endpoint selection + upstream-safety, and the
  failover/cache machinery. readHealthMetaKv is injected via configureRpcProxy
  (same wiring as configureAnalytics) to keep the import acyclic.

api.mjs stays the thin router: 5192 -> 3247 lines. All test-facing helpers are
re-exported from api.mjs so the rpc/staging/entity test suites keep importing
from workers/api.mjs unchanged.

Closes #1763
